requirements.txt
fastapi 
uvicorn 
requests 
pydantic
beautifulsoup4
google-genai
transformers
tweepy
python-dotenv
aiohttp
Brotli
snscrape
openai
nltk